requirements.txt
python-telegram-bot
playwright
httpx
apscheduler  
openai        
SQLAlchemy    
python-dotenv
google